chore: move workers to dedicated folder
